requirements.txt
aiofiles==23.1.0
aiosmtplib==2.0.1
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-mail==1.2.5
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.2
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.6
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.2.0
pypika-tortoise==0.1.6
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
toml==0.10.2
tortoise-orm==0.19.3
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
